Fix code review issues: extract shared utils, use design tokens
- Extract session history DB logic (StoredSession type, initSessionDB,
  getRecentSessions, saveSessionToHistory, formatRelativeTime,
  extractBoardName) into shared module at app/lib/session-history-db.ts
- Update session-history-panel.tsx to import from shared module and
  re-export saveSessionToHistory for backwards compat with QueueContext
- Update user-drawer.tsx to import from shared module instead of
  duplicating the IndexedDB logic
- Replace hardcoded #EF4444 in user-drawer.module.css with
  var(--color-error) CSS variable
- Add --color-error to global CSS design tokens (index.css)
- Replace inline style on avatar button with .avatarButton CSS class
- Improve error handling: catch IndexedDB failures gracefully with
  explicit fallback to empty list instead of silent console.error
